# apiDocumentation/swagger.yaml
openapi: 3.0.0
info:
  description: This is the ISE-Learning REST API
  version: 0.0.0
  title: ISE-Learning REST API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/v0/challenges/questionnaires/{questionnaireId}':
    get:
      tags:
        - Challenges
      summary: Retreives all challenges from a given questionnaire
      operationId: getAllChallengesByQuestionnaireId
      description: |
        Retrieves every challenge from a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique id.
      responses:
        '200':
          description: Challenge
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challenge'
  '/v0/challenges/{challengeId}/answers/{userId}':
    get:
      tags:
        - Challenges
      summary: Retreives a challenge answer
      operationId: getChallengeAnswer
      description: |
        Retrieves every challenge answer of the user for a specific challenge.
      parameters:
        - in: path
          name: challengeId
          schema:
            type: integer
          required: true
          description: Challenge unique id.
        - in: path
          name: userId
          schema:
            type: integer
          required: true
          description: User unique id.
      responses:
        '200':
          description: Challenge answer
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/challengeAnswer'
...
  '/v0/questionnaires/{questionnaireId}':
    get:
      tags:
        - Questionnaires
      summary: Retrieves a questionnaire
      operationId: getQuestionnaireById
      description: |
        Retrieves a single questionnaire by its unique identifer.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '200':
          description: Questionnaire
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    put:
      tags:
        - Questionnaires
      summary: Updates a questionnaire
      operationId: updateQuestionnaireById
      description: |
        Updates a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      requestBody:
        description: Questionnaire to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireModel'
      responses:
        '200':
          description: Successful updated questionnaire.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaire'
    delete:
      tags:
        - Questionnaires
      summary: Deletes a questionnaire
      operationId: deleteQuestionnaireById
      description: |
        Delete a questionnaire by its unique identifier.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '204':
          description: Successful deleted questionnaire.
          content:
            application/json:
              schema:
                type: object
...
  '/v0/questionnaireInstances/{questionnaireInstanceId}':
    get:
      tags:
        - Questionnaire Instances
      summary: Retrieves a questionnaire instance
      operationId: getQuestionnaireInstanceById
      description: |
        Retrieves a single questionnaire instance by its unique identifer.
      parameters:
        - in: path
          name: questionnaireInstanceId
          schema:
            type: integer
          required: true
          description: Questionnaire instance unique identifer.
      responses:
        '200':
          description: Questionnaire instance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireInstance'
    put:
      tags:
        - Questionnaire Instances
      summary: Updates a questionnaire instance
      operationId: updateQuestionnaireInstanceById
      description: |
        Updates a questionnaire instance by its unique identifier.
      parameters:
        - in: path
          name: questionnaireInstanceId
          schema:
            type: integer
          required: true
          description: Questionnaire instance unique identifer.
      requestBody:
        description: Questionnaire instance to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireInstanceModel'
      responses:
        '200':
          description: Successful updated questionnaire instance.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireInstance'
    delete:
      tags:
        - Questionnaire Instances
      summary: Deletes a questionnaire instance
      operationId: deleteQuestionnaireInstanceById
      description: |
        Delete a questionnaire instance by its unique identifier.
      parameters:
        - in: path
          name: questionnaireInstanceId
          schema:
            type: integer
          required: true
          description: Questionnaire instance unique identifer.
      responses:
        '204':
          description: Successful deleted questionnaire instance.
          content:
            application/json:
              schema:
                type: object
...
  '/v0/questionnaireInstances/questionnaires/{questionnaireId}':
    get:
      tags:
        - Questionnaire Instances
      summary: Retrieves a all of questionnaire instances from a questionnaire
      operationId: getAllQuestionnaireInstancesByQuestionnaireId
      description: |
        Retrieves a collection of questionnaire instances from a questionnaire by its universal unique identifer.
      parameters:
        - in: path
          name: questionnaireId
          schema:
            type: integer
          required: true
          description: Questionnaire unique identifer.
      responses:
        '200':
          description: Questionnaire's questionnaire instance collection.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireInstance'
  '/v0/questionnaireAnswer':
    post:
      tags:
        - Questionnaire Answers
      summary: Creates a questionnaire answer
      operationId: createQuestionnaireAnswer
      description: |
        Creates a questionnaire answer based on payload content.
      requestBody:
        description: Questionnaire answer to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireAnswerModel'
      responses:
        '201':
          description: Successful questionnaire answer created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireAnswer'
  '/v0/questionnaireAnswer/{questionnaireAnswerId}':
    get:
      tags:
        - Questionnaire Answers
      summary: Retrieves a questionnaire answer
      operationId: getQuestionnaireAnswerById
      description: |
        Retrieves a single questionnaire answer by its unique identifer.
      parameters:
        - in: path
          name: questionnaireAnswerId
          schema:
            type: integer
          required: true
          description: Questionnaire answer unique identifer.
      responses:
        '200':
          description: Questionnaire answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireAnswer'
    put:
      tags:
        - Questionnaire Answers
      summary: Updates a questionnaire answer
      operationId: updateQuestionnaireAnswerById
      description: |
        Updates a questionnaire answer by its unique identifier.
      parameters:
        - in: path
          name: questionnaireAnswerId
          schema:
            type: integer
          required: true
          description: Questionnaire answer unique identifer.
      requestBody:
        description: Questionnaire answer to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/questionnaireAnswerModel'
      responses:
        '200':
          description: Successful updated questionnaire answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireAnswer'
    delete:
      tags:
        - Questionnaire Answers
      summary: Deletes a questionnaire answer
      operationId: deleteQuestionnaireAnswerById
      description: |
        Delete a questionnaire answer by its unique identifier.
      parameters:
        - in: path
          name: questionnaireAnswerId
          schema:
            type: integer
          required: true
          description: Questionnaire answer unique identifer.
      responses:
        '204':
          description: Successful deleted questionnaire instance.
          content:
            application/json:
              schema:
                type: object
  '/v0/questionnaireAnswer/questionnaireInstances/{questionnaireInstanceId}':
    get:
      tags:
        - Questionnaire Answers
      summary: Retrieves all questionnaire answers from a questionnaire instance
      operationId: getAllQuestionnaireAnswersFromQuestionnaireInstanceById
      description: |
        Retrieves a collection of questionnaire answers from a questionnaire instance by its universal unique identifer.
      parameters:
        - in: path
          name: questionnaireInstanceId
          schema:
            type: integer
          required: true
          description: Questionnaire instance unique identifer.
      responses:
        '200':
          description: Questionnaire answer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  payload:
                    $ref: '#/components/schemas/questionnaireAnswer'
...
components:
  schemas:
...
    challenge:
      type: object
      required:
        - challengeID
        - challengeText
        - isPrivate
        - solutions
      properties:
        challengeID:
          type: integer
          example: 548
          description: Unique identifier of challenge.
        challengeText:
          type: string
          example: 'Implement Hello world, by printing Hello world into the console.'
          description: Test describing the challenge and its goal.
        isPrivate:
          type: boolean
          example: true
          description: >-
            Boolean value indicating whether the challenge can only be seen by
            its creator
        solutions:
          type: array
          description: List of challenges.
          items:
            $ref: '#/components/schemas/solution'
...
    solution:
      type: object
      required:
        - solutionId
        - codeLanguage
        - challengeCode
        - solutionCode
        - unitTests
      properties:
        solutionId:
          type: integer
          example: 79884
          description: Unique id to which uniquely identifies the solution.
        codeLanguage:
          type: string
          example: java
          description: Text specifying in which language the solution is.
        challengeCode:
          type: string
          example: 'public class code { public static void main(String[] args) { }}'
          description: Default code of the code presented to user.
        solutionCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests to which the the user submitted will be tested against.
...
    challengeAnswer:
      type: object
      required:
        - challengeAnswerId
        - challengeId
        - userID
        - codeLanguage
        - answerCode
        - unitTests
        - isCorrect
      properties:
        challengeAnswerId:
          type: integer
          example: 6546
          description: Unique id to which uniquely identify the challenge answer.
        challengeId:
          type: integer
          example: 55874
          description: Unique challenge id to which the answer belongs to.
        userID:
          type: integer
          example: 54954
          description: Unique user id of user to which the answer belongs to.
        codeLanguage:
          type: string
          example: Java
          description: Text identifying the programming language of the answer.
        answerCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Solution of the challenge submitted.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests developed by the user.
        isCorrect:
          type: boolean
          example: true
          description: Boolean indicating if this answer is correct or not.
...
    questionnaire:
      type: object
      required:
        - questionnaireId
        - description
        - timer
        - creatorId
        - creationDate
      properties:
        questionnaireId:
          type: integer
          example: 21313
          description: questionnaire unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
        creationDate:
          type: string
          example: '2020/05/12'
          description: questionnaire's creation date in yyyy-MM-dd format
    questionnaireModel:
      type: object
      required:
        - description
        - timer
        - creatorId
      properties:
        description:
          type: string
          example: Questionnaire for course PS-1920SV
          description: Text describing the questionnaire.
        timer:
          type: number
          example: 360000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire.
            If value is NULL has no time limite.
        creatorId:
          type: integer
          example: 44598
          description: Unique user identifier of user who created the questionnaire.
...
    questionnaireInstance:
      type: object
      required:
        - questionnaireId
        - questionnaireInstanceId
        - questionnaireInstanceUuid
        - description
        - timer
        - startTimestamp
        - endTimestamp
        - isFinished
      properties:
        questionnaireInstanceId:
          type: integer
          example: 85674
          description: Questionnaire instance unique identifier.
        questionnaireId:
          type: integer
          example: 2213
          description: Questionnaire unique identifier.
        questionnaireInstanceUuid:
          type: string
          example: 'c2c19e60-c32c-4413-84fc-369d6603d5be'
          description: Questionnaire universal unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV and class LI61N 
          description: Text describing the questionnaire instance.
        timer:
          type: number
          example: 1800000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire instance.
            If value is NULL has no time limite.
        startTimestamp:
          type: number
          example: 231231230000000
          description: >- 
            Time in milliseconds that user started solving the questionnaire instance.
        endTimestamp:
          type: number
          example: 231231231800000
          description: >-
            Time in milliseconds that user finished solving the questionnaire instance.
        isFinished:
          type: boolean
          example: false
          description: Indicates if user can still access questionnaire instance.
    questionnaireInstanceModel:
      type: object
      required:
        - questionnaireId
        - questionnaireInstanceUuid
        - description
        - timer
        - isFinished
      properties:
        questionnaireId:
          type: integer
          example: 2213
          description: Questionnaire unique identifier.
        questionnaireInstanceUuid:
          type: string
          example: 'c2c19e60-c32c-4413-84fc-369d6603d5be'
          description: Questionnaire universal unique identifier.
        description:
          type: string
          example: Questionnaire for course PS-1920SV and class LI61N 
          description: Text describing the questionnaire instance.
        timer:
          type: number
          example: 1800000
          description: >-
            Time in milliseconds which an user has to solve the questionnaire instance.
            If value is NULL has no time limite.
        isFinished:
          type: boolean
          example: true
          description: Indicates if user can still access questionnaire instance
...
    questionnaireAnswer:
      type: object
      required:
        - questionnaireAnswerId
        - questionnaireInstanceId
        - questionnaireChallengeId
        - answer
      properties:
        questionnaireAnswerId:
          type: integer
          example: 2131
          description: Unique identifier of questionnaire answer.
        questionnaireInstanceId:
          type: integer
          example: 55874
          description: Unique identifier of questionnaire instance.
        questionnaireChallengeId:
          type: integer
          example: 4467
          description: Unique identifier of questionnaire-challenge.
        answer:
          $ref: '#/components/schemas/Answer'
    questionnaireAnswerModel:
      type: object
      required:
        - questionnaireId
        - questionnaireInstanceId
        - challengeId
        - answer
      properties:
        questionnaireId:
          type: integer
          example: 2131
          description: Unique identifier of questionnaire.
        questionnaireInstanceId:
          type: integer
          example: 55874
          description: Unique identifier of questionnaire instance.
        challengeId:
          type: integer
          example: 44
          description: Unique identifier of challenge.
        answer:
          $ref: '#/components/schemas/AnswerModel'
    Answer:
      type: object
      required:
        - answerId
        - codeLanguage
        - answerCode
        - unitTests
        - isCorrect
      properties:
        questionnaireInstanceId:
          type: integer
          example: 55874
          description: Unique identifier of answer.
        codeLanguage:
          type: string
          example: "java"
          description: Text identifying the programming language name.
        answerCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Submitted answer.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests developed by the user.
        isCorrect:
          type: boolean
          example: false
          description: Indicates if submitted answer matches the solution.
    AnswerModel:
      type: object
      required:
        - codeLanguage
        - answerCode
        - unitTests
      properties:
        codeLanguage:
          type: string
          example: "java"
          description: Text identifying the programming language name
        answerCode:
          type: string
          example: >-
            public class code { public static void main(String[] args) {
            System.out.println(\"Hello World\");}}
          description: Submited answer.
        unitTests:
          type: string
          example: >-
            import org.junit.Assert;import org.junit.Test;public class unitTests
            {@Test public void TestCase1() {Assert.assertTrue(true);}}
          description: Unit tests developed by the user.